+++ lib/components/Header/Header.mdx
@@ -1,32 +1,20 @@
-import { Story, Meta, Controls, Canvas } from "@storybook/addon-docs";
-import Header from "../Header";
+import { Story, Meta, Canvas, Controls } from "@storybook/addon-docs";
+import Header from ".";
 import * as stories from "./Header.stories";
 
 <Meta of={stories} />
 
 # Header
 
-Header component for app layout. Supports app name, username and avatar as props. And links as children.
-
-When using links in the Header, use the `StyledLink` component to ensure mobile-responsiveness and consistency in the UI.
-
-At screen sizes less than 1100px the links are instead moved into a mobile menu toggled by a burger icon button to play nice with smaller devices.
+Header component for app layout. Supports app name, username and avatar as props.
 
 ## Default Header
 
 <Canvas of={stories.defaultHeader} />
 
-## Alternative Header with search
+## Header with search
 
 <Canvas of={stories.headerWithSearch} />
-
-## Header with Search and Links
-
-<Canvas of={stories.headerWithSearchAndLinks} />
-
-## Header with workspace switching
-
-<Canvas of={stories.headerWithWorkspaceSwitching} />
 
 ## Properties
