+++ lib/components/HeaderComplex/HeaderComplex.mdx
@@ -1,0 +1,33 @@
+import { Story, Meta, Controls, Canvas } from "@storybook/addon-docs";
+import HeaderComplex from "../HeaderComplex";
+import * as stories from "./HeaderComplex.stories";
+
+<Meta of={stories} />
+
+# Header
+
+Header component for app layout. Supports app name, username and avatar as props. And links as children.
+
+When using links in the Header, use the `StyledLink` component to ensure mobile-responsiveness and consistency in the UI.
+
+At screen sizes less than 1100px the links are instead moved into a mobile menu toggled by a burger icon button to play nice with smaller devices.
+
+## Default Header
+
+<Canvas of={stories.defaultHeader} />
+
+## Alternative Header with search
+
+<Canvas of={stories.headerWithSearch} />
+
+## Header with Search and Links
+
+<Canvas of={stories.headerWithSearchAndLinks} />
+
+## Header with workspace switching
+
+<Canvas of={stories.headerWithWorkspaceSwitching} />
+
+## Properties
+
+<Controls component={HeaderComplex} />
