+++ lib/index.js
@@ -32,6 +32,7 @@ export { default as Divider } from "./components/Divider";
 export { default as Expandable } from "./components/Expandable";
 export { default as Flex, FlexItem } from "./components/Flex";
 export { default as Grid, GridItem } from "./components/Grid";
+export { default as HeaderComplex } from "./components/HeaderComplex";
 export { default as Header } from "./components/Header";
 export { default as HeaderSimple } from "./components/HeaderSimple";
 export { default as Icon } from "./components/Icon";
